+++ src/games/gamebagg.h
@@ -77,7 +77,7 @@ public:
   /// @name Players
   //@{
   /// Returns the number of players in the game
-  int NumPlayers() const override { return m_players.size(); }
+  size_t NumPlayers() const override { return m_players.size(); }
   /// Returns the pl'th player in the game
   GamePlayer GetPlayer(int pl) const override { return m_players[pl]; }
   /// Returns the chance (nature) player
@@ -99,7 +99,7 @@ public:
   /// @name Outcomes
   //@{
   /// Returns the number of outcomes defined in the game
-  int NumOutcomes() const override { throw UndefinedException(); }
+  size_t NumOutcomes() const override { throw UndefinedException(); }
   /// Returns the index'th outcome defined in the game
   GameOutcome GetOutcome(int index) const override { throw UndefinedException(); }
   /// Creates a new outcome in the game
@@ -113,7 +113,7 @@ public:
   /// Returns the root node of the game
   GameNode GetRoot() const override { throw UndefinedException(); }
   /// Returns the number of nodes in the game
-  int NumNodes() const override { throw UndefinedException(); }
+  size_t NumNodes() const override { throw UndefinedException(); }
   //@}
 
   /// @name General data access
